ci: fix pre-existing CI breakage and address PR review feedback
CI/build (P4) - the red CI was pre-existing environmental rot, not the
P0 code changes. Ubuntu runners moved to 24.04 where the Aravis dev
package was renamed libaravis-0.8-dev -> libaravis-dev (still Aravis
0.8; there is no stable 0.10). Two CI actions were deleted/deprecated.

- workflows/test.yml: install libaravis-dev (not -0.8-dev); pin every
  job to Go 1.23 via a GO_VERSION env (matches go.mod); replace the
  deleted securecodewarrior gosec action with securego/gosec; bump
  checkout/setup-go/cache/upload-artifact/codecov/codeql to current
  majors; enforce formatting via gofmt -l; drop the arm64 placeholder
  no-op job; make the security scan advisory (continue-on-error)
- Dockerfile: golang:1.23-bookworm base, apt-install libaravis-dev,
  build the correct package path, drop bogus -mod=vendor and arv-tool
- Makefile: GO_VERSION 1.21 -> 1.23
- .golangci.toml: replace the copy-pasted template with a curated
  config for this library; verified clean with golangci-lint 2.5.0
- .gitignore: ignore coverage.out/coverage.html

Review feedback + related cleanups:
- bayer.go: reflect out-of-bounds neighbors across the edge instead of
  clamping, so odd-sized images keep the correct Bayer color phase
- device.go: SetNodeFeatureValue sets ARV_DEVICE_ERROR_FEATURE_NOT_FOUND
  when the feature is missing, honoring the error contract
- internal.go: remove dead fastError and its errors import
- fix unnecessary conversions flagged by unconvert (buffer/internal/
  examples/tests)
- tests/bayer_test.go: cover the edge-panic regression, opaque alpha,
  and edge-phase reflection

Verified locally against Aravis 0.8.30: go build, go vet, golangci-lint
(0 issues), and go test -race all pass.

+++ examples/performance_demo/main.go
@@ -104,7 +104,7 @@ func main() {
 	buffers := make([]aravis.Buffer, numBuffers)
 
 	for i := range numBuffers {
-		buffer, err := aravis.NewBuffer(uint(payloadSize))
+		buffer, err := aravis.NewBuffer(payloadSize)
 		if err != nil {
 			log.Fatal(err)
 		}
